requirements.txt
rich==12.0.0
pysocks==1.7.1
cloudscraper==1.2.60
